requirements.txt
asgiref
autopep8
Django==2.2.*
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
mysqlclient
django-health-check
django-debug-toolbar
IPy
boto3
